+++ src/test/java/subway/acceptance/AcceptanceTest.java
@@ -1,4 +1,4 @@
-package subway.util;
+package subway.acceptance;
 
 
 import io.restassured.RestAssured;
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
+import subway.util.DatabaseCleaner;
 
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public abstract class AcceptanceTest {
