+++ src/test/java/subway/util/StationApiRequest.java
@@ -1,9 +1,10 @@
-package subway;
+package subway.util;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import org.springframework.http.MediaType;
+import subway.dto.StationRequest;
 
 public class StationApiRequest {
